Change  on indicator lib

+++ ESP32/Load_Demand_Response_System/code/lib/Display/led_indicator/led_indicator.h
@@ -2,6 +2,6 @@
 
 namespace led_indicator
 {
-  void begin(int pin);
+  void begin(uint8_t pin);
   void update();
 }
